route操作系の処理をlibraryに移動
